box: remove dead code from box_process_call/eval()
box_process_call/eval() in the end check if there is an
active transaction. If there is, it is rolled back, and
an error is set.

But rollback is not needed anymore, because anyway in
the end of the request the fiber is stopped, and its
not finished transaction is rolled back. Just setting
of the error is enough.

Follow-up #4662

(cherry picked from commit f5d514485e583fdac94b9f003cb9f0c6529e4350)