wal: warn when trying to write a record with a broken lsn
There is an assertion in vclock_follow `lsn > prev_lsn`, which doesn't
fire in release builds, of course. Let's at least warn the user on an
attempt to write a record with a duplicate or otherwise broken lsn, and
not follow such an lsn.

Follow-up #4739

(cherry picked from commit e0750262b899506649fb65592ce645129edc2687)