session: deprecate box.session.push() 'sync'
box.session.push() has two parameters - data to push and 'sync'.
The sync is a request ID with which the out of bound data should
be pushed into a socket.

This was introduced as a workaround for #3450, and is useless
since its resolution.

A user anyway can't push to different sessions, where that
parameter could be useful. And pushing into requests of the same
session, on the contrary, is something not really needed anywhere,
not portable to non-binary session types (console, background),
and is just dangerous since it is easy to add a bug here.

The patch makes the parameter deprecated. Now on its usage there
will be a rate limited warning in the logs.

Part of #4689