test: update test-run
It is hotfix for the previous test-run update, where 'core = app' test
suites starts to support 'use_unix_sockets_iproto' option. We found a
case when unix socket names may clash and attempt to listen on the
socket will report 'Address already in use' error. It is fixed in this
test-run update.

See more in [1].

[1]: https://github.com/tarantool/test-run/pull/211